Update internal documentation and update package layout
Since the README file is not enough I updated the javadoc comments
to reflect the ideas behind the implementation and re-organized
the package layout.
