+++ .editorconfig
@@ -6,7 +6,7 @@ end_of_line = lf
 indent_style = space
 indent_size = 4
 insert_final_newline = true
+max_line_length = 80
 
 [*.{java,groovy,gradle}]
 trim_trailing_whitespace = true
-max_line_length = 120
