+++ src/main/java/com/jossemargt/cookietwist/exception/InvalidFormatException.java
@@ -23,21 +23,49 @@
  */
 package com.jossemargt.cookietwist.exception;
 
+/**
+ * The Class InvalidFormatException is meant to describe any CookieValue Model
+ * de-serialization fault, mostly related to String formatting, numeric parsing
+ * or un-expected amount of String tokens.
+ */
 @SuppressWarnings("serial")
 public class InvalidFormatException extends Exception {
 
+    /**
+     * Instantiates a new invalid format exception without message string or
+     * cause.
+     */
     public InvalidFormatException() {
         super();
     }
 
+    /**
+     * Instantiates a new invalid format exception with a string message
+     * describing the occurrence.
+     *
+     * @param message the description of the occurrence.
+     */
     public InvalidFormatException(String message) {
         super(message);
     }
 
+    /**
+     * Instantiates a new invalid format exception with a string message
+     * describing the occurrence and the Exception that triggered it.
+     *
+     * @param message the description of the occurrence.
+     * @param cause the Exception which triggerd this one.
+     */
     public InvalidFormatException(String message, Throwable cause) {
         super(message, cause);
     }
 
+    /**
+     * Instantiates a new invalid format exception whith the one that triggered
+     * it.
+     *
+     * @param cause the Exception which triggerd this one.
+     */
     public InvalidFormatException(Throwable cause) {
         super(cause);
     }
